# internal/library/handlers.go
package library

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"os"
	"path/filepath"
	"strconv"
	"strings"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/viewdock/viewdock/internal/audit"
	"github.com/viewdock/viewdock/internal/auth"
	"github.com/viewdock/viewdock/internal/httpapi"
)

func (s *Service) Routes(r chi.Router) {
	r.Get("/libraries", s.handleListLibraries)
	r.Post("/libraries", s.handleCreateLibrary)
	r.Get("/libraries/{id}", s.handleGetLibrary)
	r.Patch("/libraries/{id}", s.handlePatchLibrary)
	r.Delete("/libraries/{id}", s.handleDeleteLibrary)
	r.Post("/libraries/{id}/scan", s.handleScanLibrary)
	s.moveRoutes(r)

	r.Get("/movies", s.handleListMovies)
	r.Get("/movies/{id}", s.handleGetMovie)
	r.Get("/browse/signals", s.handleBrowseSignals)
	s.preferenceRoutes(r)
	r.Get("/series", s.handleListSeries)
	r.Get("/series/{id}", s.handleGetSeries)
	r.Get("/series/{id}/next", s.handleNextEpisode)
	r.Get("/next-up", s.handleNextUp)
	r.Get("/episodes/{id}", s.handleGetEpisode)
	r.Get("/artwork/{kind}/{itemKind}/{id}", s.handleArtwork)
	r.Put("/movies/{id}/rating", s.handleSetRating("movie"))
	r.Delete("/movies/{id}/rating", s.handleResetRating("movie"))
	r.Put("/series/{id}/rating", s.handleSetRating("series"))
	r.Delete("/series/{id}/rating", s.handleResetRating("series"))
}

func (s *Service) handleSetRating(itemKind string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		p := auth.FromRequest(r)
		if !p.HasPerm(auth.PermLibrariesManage) {
			httpapi.WriteErr(w, http.StatusForbidden, "forbidden", "permission required")
			return
		}
		var body struct {
			ContentRating string `json:"content_rating"`
			RatingAge     *int   `json:"rating_age"`
		}
		if err := httpapi.ReadJSON(r, &body); err != nil {
			httpapi.WriteErr(w, http.StatusBadRequest, "bad_request", "invalid json")
			return
		}
		rating := strings.TrimSpace(body.ContentRating)
		if len(rating) > 32 {
			httpapi.WriteErr(w, http.StatusBadRequest, "rating", "content_rating is too long")
			return
		}
		age := body.RatingAge
		if age == nil && rating != "" {
			mapped, ok := CertificationAge("", rating)
			if !ok {
				httpapi.WriteErr(w, http.StatusBadRequest, "rating", "unknown rating; provide rating_age")
				return
			}
			age = &mapped
		}
		id := chi.URLParam(r, "id")
		if err := s.SetRating(r.Context(), itemKind, id, rating, age); err != nil {
			writeLibErr(w, err)
			return
		}
		detail := "rating=" + rating + " age=unrated"
		if age != nil {
			detail = "rating=" + rating + " age=" + strconv.Itoa(*age)
		}
		s.audit(r, p.UserID, "content.rating.set", itemKind+"/"+id, detail)
		httpapi.WriteJSON(w, http.StatusOK, map[string]any{"content_rating": rating, "rating_age": age, "rating_source": "admin"})
	}
}

func (s *Service) handleResetRating(itemKind string) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		p := auth.FromRequest(r)
		if !p.HasPerm(auth.PermLibrariesManage) {
			httpapi.WriteErr(w, http.StatusForbidden, "forbidden", "permission required")
			return
		}
		id := chi.URLParam(r, "id")
		if err := s.ResetRating(r.Context(), itemKind, id); err != nil {
			writeLibErr(w, err)
			return
		}
		s.audit(r, p.UserID, "content.rating.reset", itemKind+"/"+id, "")
		httpapi.WriteOK(w)
	}
}

func (s *Service) audit(r *http.Request, actorID, action, target, detail string) {
	log := s.Audit
	if log == nil {
		log = audit.New(s.DB)
	}
	log.Event(r.Context(), actorID, action, target, httpapi.ClientIPString(r, s.Cfg), detail)
}

func writeListErr(w http.ResponseWriter, err error) {
	if errors.Is(err, ErrUnavailable) {
		httpapi.WriteErr(w, http.StatusServiceUnavailable, "catalogue_unavailable", ErrUnavailable.Error())
		return
	}
	slog.Error("catalogue listing failed", "category", "library", "err", err)
	httpapi.WriteErr(w, http.StatusInternalServerError, "library", "the catalogue could not be loaded")
}

func (s *Service) handleListLibraries(w http.ResponseWriter, r *http.Request) {
	libs, err := s.List(r.Context())
	if err != nil {
		httpapi.WriteErr(w, 500, "library", err.Error())
		return
	}
	if ids := grantedFilter(r.Context(), nil); ids != nil {
		allow := map[string]bool{}
		for _, id := range ids {
			allow[id] = true
		}
		filtered := libs[:0]
		for _, lib := range libs {
			if allow[lib.ID] {
				filtered = append(filtered, lib)
			}
		}
		libs = filtered
	}
	if libs == nil {
		libs = []Library{}
	}
	for i := range libs {
		libs[i] = s.withHostPath(libs[i])
	}
	httpapi.WriteJSON(w, 200, libs)
}

func (s *Service) handleCreateLibrary(w http.ResponseWriter, r *http.Request) {
	var body struct {
		Name        string `json:"name"`
		Path        string `json:"path"`
		ContentType string `json:"content_type"`
	}
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		httpapi.WriteErr(w, 400, "bad_request", "invalid json")
		return
	}
	lib, err := s.Create(r.Context(), body.Name, body.Path, body.ContentType)
	if err != nil {
		httpapi.WriteErr(w, 400, "library", err.Error())
		return
	}
	httpapi.WriteJSON(w, 201, s.withHostPath(lib))
}

func (s *Service) handleGetLibrary(w http.ResponseWriter, r *http.Request) {
	lib, err := s.Get(r.Context(), chi.URLParam(r, "id"))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, s.withHostPath(lib))
}

func (s *Service) handlePatchLibrary(w http.ResponseWriter, r *http.Request) {
	var body struct {
		Name           *string `json:"name"`
		Path           *string `json:"path"`
		ContentType    *string `json:"content_type"`
		UploadsEnabled *bool   `json:"uploads_enabled"`
	}
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		httpapi.WriteErr(w, 400, "bad_request", "invalid json")
		return
	}
	lib, err := s.Update(r.Context(), chi.URLParam(r, "id"), Patch{
		Name: body.Name, RootPath: body.Path, ContentType: body.ContentType, UploadsEnabled: body.UploadsEnabled,
	})
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, s.withHostPath(lib))
}

func (s *Service) handleDeleteLibrary(w http.ResponseWriter, r *http.Request) {
	if err := s.Delete(r.Context(), chi.URLParam(r, "id")); err != nil {
		writeLibErr(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (s *Service) handleScanLibrary(w http.ResponseWriter, r *http.Request) {
	id, err := s.StartScan(r.Context(), chi.URLParam(r, "id"))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 202, map[string]string{"scan_run_id": id})
}

func (s *Service) handleListMovies(w http.ResponseWriter, r *http.Request) {
	list, err := s.ListMovies(r.Context(), GrantedIDsFrom(r.Context()))
	if err != nil {
		writeListErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, list)
}

func (s *Service) handleGetMovie(w http.ResponseWriter, r *http.Request) {
	m, err := s.GetMovie(r.Context(), chi.URLParam(r, "id"))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, m)
}

func (s *Service) handleListSeries(w http.ResponseWriter, r *http.Request) {
	list, err := s.ListSeries(r.Context(), GrantedIDsFrom(r.Context()))
	if err != nil {
		writeListErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, list)
}

func (s *Service) handleGetSeries(w http.ResponseWriter, r *http.Request) {
	ser, err := s.GetSeries(r.Context(), chi.URLParam(r, "id"))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, ser)
}

func (s *Service) handleNextEpisode(w http.ResponseWriter, r *http.Request) {
	ep, err := s.NextEpisode(r.Context(), chi.URLParam(r, "id"), UserIDFrom(r.Context()))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, ep)
}

func (s *Service) handleGetEpisode(w http.ResponseWriter, r *http.Request) {
	ep, err := s.GetEpisode(r.Context(), chi.URLParam(r, "id"))
	if err != nil {
		writeLibErr(w, err)
		return
	}
	httpapi.WriteJSON(w, 200, ep)
}

func (s *Service) handleArtwork(w http.ResponseWriter, r *http.Request) {
	kind := chi.URLParam(r, "kind")
	itemKind := chi.URLParam(r, "itemKind")
	id := chi.URLParam(r, "id")
	if !validArtworkKind(kind) || !validItemKind(itemKind) {
		httpapi.WriteErr(w, 404, "not_found", "not found")
		return
	}
	if err := s.visible(r.Context(), itemKind, id); err != nil {
		writeLibErr(w, err)
		return
	}
	var stored string
	_ = s.DB.QueryRowContext(r.Context(), `
		SELECT path FROM artwork WHERE item_kind = ? AND item_id = ? AND kind = ?
	`, itemKind, id, kind).Scan(&stored)

	path := ""
	if stored != "" {
		if filepath.IsAbs(stored) {
			path = stored
		} else if s.CacheDir != "" {
			path = filepath.Join(s.CacheDir, stored)
		}
	}
	if path == "" && s.CacheDir != "" {
		for _, ext := range []string{".jpg", ".jpeg", ".png", ".webp"} {
			cand := filepath.Join(s.CacheDir, "artwork", kind, itemKind, id+ext)
			if st, err := os.Stat(cand); err == nil && st.Mode().IsRegular() {
				path = cand
				break
			}
		}
	}
	if path == "" || os.IsNotExist(mustStat(path)) {
		if kind == "thumb" && s.Thumber != nil {
			if p, err := s.lazyThumb(r, itemKind, id); err == nil && p != "" {
				path = p
			}
		}
	}
	if path == "" {
		httpapi.WriteErr(w, 404, "not_found", "not found")
		return
	}
	if s.CacheDir != "" {
		if err := ContainsPath(s.CacheDir, path); err != nil {
			httpapi.WriteErr(w, 404, "not_found", "not found")
			return
		}
	}
	f, err := os.Open(path)
	if err != nil {
		httpapi.WriteErr(w, 404, "not_found", "not found")
		return
	}
	defer f.Close()
	st, err := f.Stat()
	if err != nil || !st.Mode().IsRegular() {
		httpapi.WriteErr(w, 404, "not_found", "not found")
		return
	}
	ctype := "image/jpeg"
	switch strings.ToLower(filepath.Ext(path)) {
	case ".png":
		ctype = "image/png"
	case ".webp":
		ctype = "image/webp"
	}
	w.Header().Set("Content-Type", ctype)
	w.Header().Set("Cache-Control", "private, max-age=86400")
	http.ServeContent(w, r, st.Name(), st.ModTime(), f)
}

func (s *Service) lazyThumb(r *http.Request, itemKind, id string) (string, error) {
	if s.CacheDir == "" || s.Thumber == nil {
		return "", errors.New("no thumber")
	}
	loc, err := s.LocateItem(r.Context(), itemKind, id)
	if err != nil || loc == nil {
		return "", ErrNotFound
	}
	if err := s.Contains(loc.LibraryID, loc.AbsPath); err != nil {
		return "", err
	}
	destDir := filepath.Join(s.CacheDir, "artwork", "thumb", itemKind)
	if err := os.MkdirAll(destDir, 0o755); err != nil {
		return "", err
	}
	dest := filepath.Join(destDir, id+".jpg")
	at := loc.DurationMS / 10
	if at < 1000 {
		at = 5000
	}
	if err := s.Thumber.Thumb(r.Context(), loc.AbsPath, dest, at); err != nil {
		return "", err
	}
	rel := filepath.ToSlash(filepath.Join("artwork", "thumb", itemKind, id+".jpg"))
	_, _ = s.DB.ExecContext(r.Context(), `
		INSERT INTO artwork(id, item_kind, item_id, kind, path, source, locked)
		VALUES (?, ?, ?, 'thumb', ?, 'thumb', 0)
		ON CONFLICT(item_kind, item_id, kind) DO UPDATE SET path = excluded.path
	`, uuid.NewString(), itemKind, id, rel)
	return dest, nil
}

func writeLibErr(w http.ResponseWriter, err error) {
	if errors.Is(err, ErrNotFound) {
		httpapi.WriteErr(w, 404, "not_found", "not found")
		return
	}
	if errors.Is(err, ErrUnavailable) {
		httpapi.WriteErr(w, http.StatusServiceUnavailable, "catalogue_unavailable", ErrUnavailable.Error())
		return
	}
	httpapi.WriteErr(w, 400, "library", err.Error())
}

func validArtworkKind(k string) bool {
	return k == "poster" || k == "backdrop" || k == "thumb"
}

func validItemKind(k string) bool {
	return k == "movie" || k == "series" || k == "episode"
}

func mustStat(path string) error {
	if path == "" {
		return os.ErrNotExist
	}
	_, err := os.Stat(path)
	return err
}