# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ViewDock API
  version: 0.1.0
  license:
    name: AGPL-3.0-or-later
servers:
  - url: /api/v1
paths:
...
  /libraries/{id}:
    get:
      operationId: getLibrary
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
    patch:
      operationId: patchLibrary
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
    delete:
      operationId: deleteLibrary
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204": { description: Deleted }
  /libraries/{id}/scan:
    post:
      operationId: scanLibrary
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "202": { description: Started }
  /library-moves/preview:
    post:
      operationId: previewLibraryMove
      description: Plans moving titles to another library without changing anything. Needs libraries.manage.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MoveRequest" }
      responses:
        "200": { description: Eligible and ineligible titles with reasons }
        "404": { description: Library not found }
  /library-moves:
    post:
      operationId: startLibraryMove
      description: Moves the eligible titles on disk and in the catalogue in the background. Needs libraries.manage.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MoveRequest" }
      responses:
        "202": { description: Move job started }
        "409": { description: Another move or a scan of one of the libraries is running }
  /library-moves/{id}:
    get:
      operationId: getLibraryMove
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: Move job with the outcome of each title }
        "404": { description: Not found }
...
  /movies/{id}:
    get:
      operationId: getMovie
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
...
  /series/{id}:
    get:
      operationId: getSeries
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
  /series/{id}/next:
    get:
      operationId: nextEpisode
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
  /episodes/{id}:
    get:
      operationId: getEpisode
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
...
  /uploads:
    get:
      operationId: listUploads
      responses:
        "200": { description: Admin upload sessions }
    post:
      operationId: createUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUploadRequest"
      responses:
        "201": { description: Created }
  /uploads/{id}:
    get:
      operationId: getUpload
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: Upload session }
    head:
      operationId: uploadHead
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: OK }
    put:
      operationId: uploadPut
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: Bytes accepted }
        "409": { description: Offset conflict }
    delete:
      operationId: uploadDelete
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204": { description: Cancelled }
...
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    MoveRequest:
      type: object
      required: [destination_library_id]
      properties:
        source_library_id: { type: string, description: "With all: move every title of this library" }
        destination_library_id: { type: string }
        all: { type: boolean }
        items:
          type: array
          items:
            type: object
            required: [kind, id]
            properties:
              kind: { type: string, enum: [movie, series] }
              id: { type: string }
...
    CreateUploadRequest:
      type: object
      required: [library_id, filename]
      properties:
        library_id: { type: string }
        filename: { type: string }
        size: { type: integer, format: int64 }
        size_bytes: { type: integer, format: int64 }
        mime: { type: string }
...